feat: Reading User input challenge - v2

+++ section6/Reading User Input Challenge/src/Main.java
@@ -5,13 +5,13 @@ public class Main {
         Scanner scanner = new Scanner(System.in);
 
         int counter = 1;
-        int sum = 0;
+        double sum = 0;
 
         while (counter <= 5) {
             System.out.println("Enter number #" + counter + ":");
             String nextNumber = scanner.nextLine();
             try {
-                double number = Integer.parseInt(nextNumber);
+                double number = Double.parseDouble(nextNumber);
                 counter++;
                 sum += number;
             } catch (NumberFormatException e) {
